Update register-task-with-maintenance-window.html
"instanceId" in the 3 code examples on this page will cause a problem.  It will not prevent the task from being registered with the window, but when the task runs during the maintenance window, it will fail due to error "The supplied parameters for invoking the specified Automation document are incorrect."  In my testing if replaced with "InstanceId" the task will execute successfully.